Move useLastMetricsReportSentAt to its own file
